refact : 웹스크래핑 구조 개선
- 웹스크래핑 함수에서 평가도 같이 진행
- index 실행 시 한번만 웹스크래핑 하도록 변경
